Build with msys2
Fix errors, reported by gcc 12.1.0 on MSYS2. Untie from old borland's
proprietary resource compiler.

Only 32-bit build is still supported; 64-bit one needs more cleanup.

+++ mediaaccess.cpp
@@ -25,7 +25,7 @@ cDiskAccess::cDiskAccess(char num)
   }
 
   if(!
-    DeviceIoControl(hDisk, IOCTL_DISK_GET_DRIVE_GEOMETRY, NULL, 0, &diskGeometry, sizeof(DISK_GEOMETRY), &(DWORD)br, NULL)
+    DeviceIoControl(hDisk, IOCTL_DISK_GET_DRIVE_GEOMETRY, NULL, 0, &diskGeometry, sizeof(DISK_GEOMETRY), &br, NULL)
   )
   {
     iErrorCode=HEX_MEDIA_ERR_ACCESSING_DISK_GEOMETRY;
